Allows user-configured http.Client to be passed into NewTwilioClient.
Users may setup their http.Client with various useful
parameters that are not part of the http.DefaultClient  (importantly
including a Timeout). Furthermore, because http.Clients are designed
to be reused and are safe across goroutines, sharing a premade
http.Client can be beneficial.

If nil is passed in place of an *http.Client, the http.DefaultClient
is used.