Add a new GetMessage function
GetSMS only returns the SMS data (from `/SMS/Messages/{SID}.json`).
GetMessage, however, returns more of the status information (from
the `/Messages/{SID}.json`).  This is necessary in order to determine
if the message was delivered, if the handset is unavailable, if the
person has unsubscribed, etc.

In addition, I added a few missing SMS fields as part of my troubleshooting.
The GetSMS response now includes:

* `messaging_service_sid`
* `num_segments`
* `price_unit`